Expose rip_key_B in System Extra Modal
This expose the config key need to decrypt and encrypt the config file. This file is unique and random generated by the modem on the first start. This can be usefull to unlock the firmware again by having this key and a config file. As a decrypt/config utility is provided.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>