Fix #336
portforward mapper expose an ip option that is not actually present in the firewall config file. This cause both mac and ip address to show in the table. And on next edit, the ip address will take priority to the mac option. To solve this we remove the ip address option if mac option is set.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>