Fix xdslctl mapper for Us Band
The script try to handle strange situation when the DSLAM disable the first Us band if the user is near the DSLAM. This caused some problem when some Isp use only 3 band for the Us. Rework this to actually check if the first Us band is disabled and skip if not.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>