upgrade: try to handle #469 problems
On low mem device the rootdevice will try to remove dev gui and also will convert any gui to a slim version with no additional module (for example specific module for other modem)
This should reduce the size of the zip to 2-3mb. This needs to be tested but i should have updated all the script that use the gui to restore something.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>