webui: fix login for non admin user
The ajax getcard was permitted only for admin and engineer role. This forbid, for example, guest role to use the ajax and load card. Also check if the role has access to ajax and fallback to the old mode. Also set the config and rules table global to not recall the config iterator every time.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>