lib: adds helper to find the right logger to use and convert all to use it
Currently there are 2 module to log data tch.logger and transformer.logger
tch.logger is present only in recent system. This add a new function to check if tch.logger actually exist and fallback to transformer.logger if not. Convert all script to use this function instead of call the module directly and fail to run if it doesn't exist.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>