upgrade: adds rootonly option to sysupgrade
This is usefull to test new firmware. This mode root the modem but doesn't install the custom web interface, modify the config options or install custom app. Doing this should never put the modem in a bootloop state. Pay attention that the default update mode is still the reinstall gui on update.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>