web: fix tod script exception
wifitod works by saving the orig state in the config. If for some reason this config disappear in the time the rule start and when the rule should stop, nothing is done and the script exit. Fix this by invert the state set on config on the stop action of the script.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>