web: fix broken wifi tod page and reset timer on rule modify
Previously on rule modify the timers were not reseted if they were started. This caused problem with rule modification in case someone invert a rule to switch off the wifi instead of power them on. As the timer already started, the timer slot was detected and on tod restart the script skip any execution till the timer ends. Delete rule on any rule modify to reinitialize the timer and run the rule immediately.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>