mapper: fix a real bug in rpc.network.interface mapper
The static_dns_servers set was completly broken due to the fact that uci_helper.commit want the binding and NOT the config name to commit, as the real function call the config entry in the binding by itself. This is present in normal Technicolor source so....
Also show static_dns even if the interface is not a wan interface.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>