gui: introduce flag to fix any crash related to xdsl start
Use a file flag to check if for whatever reason the xdsl start cause
some kernel panic. The flag will restore the original bin if a crash is
detected. (Before xdsl start the flag is created, if the system crash
the flag is never removed and at the next boot the file is restored)

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>